third_party/WebKit/LayoutTests/imported/web-platform-tests/html/obsolete/requirements-for-implementations/other-elements-attributes-and-apis/document-all-expected.txt

This is a testharness.js-based test.
FAIL 'unusual behaviors' of document.all assert_equals: expected "undefined" but got "object"
FAIL 'unusual behaviors' of document.all with assignment assert_equals: expected "undefined" but got "object"
Harness: the test ran to completion.
